% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scanpy.R
\docType{package}
\name{ggThemeScanpy}
\alias{ggThemeScanpy}
\title{ggThemeScanpy: A package of theme modifications for ggplot that produce plots that are very similar to the default scanpy plotting options}
\section{Functions}{

All functions take a ggplot plot object and return a modified ggplot plot object.

theme_scanpy_scatter(plot.object) produces plots that are similar to scanpy PCA or UMAP scatter plots
}

