% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scanpy.R
\name{theme_scanpy}
\alias{theme_scanpy}
\title{Return a ggplot2 theme object blanking unused theme elements}
\usage{
theme_scanpy()
}
\value{
ggplot2 theme
}
\description{
Return a ggplot2 theme object blanking unused theme elements
}
