% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scanpy.R
\name{theme_scanpy_continuous}
\alias{theme_scanpy_continuous}
\title{Apply scanpy theme to a ggplot2 plot object that is continuous on both axes}
\usage{
theme_scanpy_continuous(plot.object)
}
\arguments{
\item{plot.object}{ggplot2 Plot Object}
}
\value{
ggplot2 Plot object
}
\description{
Apply scanpy theme to a ggplot2 plot object that is continuous on both axes
}
