% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scanpy.R
\name{theme_scanpy_scatter}
\alias{theme_scanpy_scatter}
\title{Apply scanpy scatter (e.g. PCA or UMAP) theme to a ggplot2 plot object}
\usage{
theme_scanpy_scatter(plot.object, color.level = "color")
}
\arguments{
\item{plot.object}{ggplot2 Plot Object}

\item{color.level}{Applies hue to either "color" or "fill" properties. Set to NULL to skip applying hue.}
}
\value{
ggplot2 Plot object
}
\description{
Apply scanpy scatter (e.g. PCA or UMAP) theme to a ggplot2 plot object
}
