% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scanpy.R
\name{theme_scanpy_x_categorical}
\alias{theme_scanpy_x_categorical}
\title{Apply scanpy theme to a ggplot2 plot object that is continuous on the y-axis and discrete on the x-axis}
\usage{
theme_scanpy_x_categorical(plot.object)
}
\arguments{
\item{plot.object}{ggplot2 Plot Object}
}
\value{
ggplot2 Plot object
}
\description{
Apply scanpy theme to a ggplot2 plot object that is continuous on the y-axis and discrete on the x-axis
}
