% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scanpy.R
\name{theme_scanpy_x_y_categorical}
\alias{theme_scanpy_x_y_categorical}
\title{Apply scanpy theme to a ggplot2 plot object that is discrete on both the x-axis and the y-axis}
\usage{
theme_scanpy_x_y_categorical(plot.object)
}
\arguments{
\item{plot.object}{ggplot2 Plot Object}
}
\value{
ggplot2 Plot object
}
\description{
Apply scanpy theme to a ggplot2 plot object that is discrete on both the x-axis and the y-axis
}
